+++ sbroad/sbroad-core/src/ir/tree/subtree.rs
@@ -296,7 +296,8 @@ fn subtree_next<'plan>(
                 Expression::Case { .. } => iter.handle_case_iter(expr),
                 Expression::Bool { .. }
                 | Expression::Arithmetic { .. }
-                | Expression::Concat { .. } => iter.handle_left_right_children(expr),
+                | Expression::Concat { .. }
+                | Expression::Index { .. } => iter.handle_left_right_children(expr),
                 Expression::Trim { .. } => iter.handle_trim(expr),
                 Expression::Like { .. } => iter.handle_like(expr),
                 Expression::Row(Row { list, .. })